Split modules by concern
create_content.py was holding five unrelated concerns. Now:

- model.py: the Person and Alias types
- aliases.py: aliases file IO (get_aliases, dump_normalized_aliases)
- git.py: shortlog subprocess and parsing into Persons
- create_content.py: content rendering only
- api.py: create_contributors_txt, the public API, re-exported
  unchanged from the package root

__main__.py only contains the command line entry point now, which also
fixes the RuntimeWarning runpy emitted on 'python -m contributors_txt'
because __init__.py imported __main__.
