+++ contributors_txt/__init__.py
@@ -3,4 +3,4 @@
 
 __all__ = ["create_contributors_txt"]
 
-from contributors_txt.__main__ import create_contributors_txt
+from contributors_txt.api import create_contributors_txt
